Move test directory out of src
